backend/requirements-gpu.txt
Django>=4.0,<5.0
djangorestframework>=3.12.0
djangorestframework-simplejwt
django-cors-headers
mysqlclient
python-decouple
pyjwt
pyshark
pandas>=1.1.0
scikit-learn>=1.0.0
celery==5.3.6
redis==5.0.1
shap
tensorflow-gpu==2.11.0
